Check for Existing Systemd dirs Before Failing
Fix the current methodology for handling systemd directories.
Currently the assumption is that all system and user-supplied
systemd directories will always be present.  This is not true
and the code resulted in an exception if a directory was missing.
This change allows any of the directories to be handled, but
doesn't require them all to be present.

(transferred from ManageIQ/manageiq-gems-pending@fc13c8c17b0ea2cb138614c2b4188d289bd3c3f1)