Remove Unused getParentPathMac Method
The getParentPathMac method has been commented out since the beginning
of this file supporting legacy Windows Virtual Disks.  Removing it now
as well as the commented out call.

In addition the use of the "MacX" case of the same method is not used, and
should be removed, so the getParentPathMacX method will also be removed,
leaving only one functional (or non-deprecated) case.  The entire case
statement will be replaced with an if-else.
In case there is no parent initialized after finishing the loop, the previously
commented-out exception will be utilized.

(transferred from ManageIQ/manageiq-gems-pending@8f8874edfdc8efa752044e70792323947e19c962)