WinRMv2 Support
Add support for the new V2 version of the WinRM Gem.
This Gem uses the PSRP protocol to communicate to PowerShell
processes running on the Windows server.

Except for elevated processes needed to access network shares (see below)
this support allows powershell to run across invocations of WinRM.  Variables
created in one invocation can continue to be accessed in further invocations
using the same shell.

The API has changed somewhat requiring changes in the Microsoft Provider code
as well as in the gems/pending code for SCVMM/HyperV SSA.

Since the WinRM V2 support is not yet finalized (but is in its final stages)
this will be added to a WIP PR for now.

(transferred from ManageIQ/manageiq-gems-pending@326a24e433ec98ca095548bfcc59da25fef0332a)