Revert Constant name to MIN_SECTORS_PER_ENTRY
The previous commit changed the name of MIN_SECTORS_PER_ENTRY
to MIN_SECTORS_TO_CACHE, in an effort to be consistent with the
MiqHyperVDisk class from which this code is extracted.  However
the original name here is more appropriate and it will be changed
back.

(transferred from ManageIQ/manageiq-gems-pending@19c69b8270a906b0b8b27596c9feeb1342dfc60b)