Upgrade to WinRM 2.1 / winrm-elevated 1.0.1
Since adding this PR as WIP WinRM has had several updates bringing it to 2.1.
Update the Gemfile to support 2.1 as well as winrm-elevated 1.0.1.

Changes to other files are part of bringing this along from a WIP to a ready-for-release
PR.

(transferred from ManageIQ/manageiq-gems-pending@8cae7c282202c79fe6aec89e843ac44949d92db5)