Fix Directory Name Length for Superblock Type 5
When an XFS filesystem with Superblock type 5 is encountered,
the spec indicates that the filename in a directory entry is to
be incremented by one to find the next field in the entry.  This
increment was occurring before capturing the true length of the name
resulting in an extra character in the name, which is obviously incorrect.
It appears a Superblock of this type had not been encountered previously.

(transferred from ManageIQ/manageiq-gems-pending@7f843284bf65ff071b947210935b99f830c52890)