Remove oVirt provider code
Currently the part of the oVirt provider that performs event monitoring
and inventory collection (using version 3 of the oVirt API and the
'ovirt' gem) lives in this repository, in the
'lib/gems/pending/ovirt_provider' directory. As the oVirt provider has
been extracted to a separate repository, it is convenient to move that
code there. The relevant copy has been added to that repository in this
pull request:

  Move code from manageiq-gems-pending
  https://github.com/ManageIQ/manageiq-providers-ovirt/pull/9

This patch removes the code from this repository.

Note that part of the code can't be moved, or rather, isn't convinient
to move it now. The reason is that code, the 'rhevmTest.rb' file, for
example, doesn't conceptually belong into the oVirt provider. To keep
this code in this repository it is necessary to add a dependency, so
that 'manageiq-gems-pending' will now depend on
'manageiq-providers-ovirt'.

https://github.com/ManageIQ/manageiq-providers-ovirt/issues/8

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

(transferred from ManageIQ/manageiq-gems-pending@d7bb7b638554cd8ca976450fd694e9d39d18696a)