SSA for Managed Images
Add support for SSA requests on Azure Managed Images.
Note that these Managed Images may or may not have Managed Disks.
In the case of an image with a Managed Disk we will get a disk name,
while in the case of a Managed Image without a Managed Disk we will
receive a URI.  For Managed Images with Managed Disks we will need
to call AzureManagedDisk passing in the Azure::Armrest::Storage::Diskservice
instead of the SnapshotService.

This is in support of blocker BZ https://bugzilla.redhat.com/show_bug.cgi?id=1496523.

There will be a parallel PR for the Azure Provider Repo.