Use the Smartstate Snapshot Name Built by the Provider
Since there is an 80 character limit on the length of the snapshot name
constructed for managed disks use the name set by the Azure provider.
If there is no name passed in the provider is using the previous code
that built the name known here and use that name instead.