Assume C: Drive if Disk Sig Doesn't Match Serial Number
If the Disk Signature of the root volume didn't match the serial number
for one of the drives found in the BCD registry (or the boot.ini previously)
the assumption was that we could not map the root to a mount point.

However - the logic in the code already made the assumption that if there were
no drives found in the registry that the C drive should be used.

This change also makes that assumption in the case that the disk signature
doesn't match a registry entry.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1518347