Check Drivetype for Each Differencing VhdxDisk
When running Smartstate Analysis on SCVMM VhdxDisks first we need to
determine if the virtual disk file lives on a local disk or on a network disk.
Accessing the network disk must be through an elevated WinRM call, while that is
not necessary for local disks.

Previously once the determination is made on the initial disk opened the same type
is assumed for each differencing (parent) disk in the chain of snapshots for a VhdxDisk.

This, of course, may be incorrect since not each disk in the chain is necessarily on either
a local or network disk.  If the initial disk is local, and the parent is remote, this would
cause an access error trying to read the parent disk.

We are changing this to check the type of disk for each new one opened.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1547766