Convert MiqS3Session to MiqS3Storage
Converts MiqS3Session to an MiqObjectStorage subclass, MiqS3Storage.

A decent amount here didn't change, but it updates the `#add` and
`#download` methods to use the MiqFileStorage::Interface and instead
override `#upload_single` and `#download_single`.

Many of the tests didn't make it over since a lot of it was in regards
to "mounting", which ObjectStorage classes don't do and is now obsolete,
as well as cleaned up MiqGenericMountSession from any references that
were s3 specific.

Also dried up the code a bit from what was implemented in MiqS3Session.

Regarding the elephant in the room...
-------------------------------------

Yes, I backported Aws::S3::MultipartStreamUploader from the v3 portion
of the aws-sdk as part of this commit.  Some form of "upload streaming"
was necessary for file splitting to work.

I could have implemented something myself with probably a little less
code, but what was already part of the v3 API, is probably tested
extensively, and works with the v2 Seahorse client and lib made more
sense to implement via this way instead of trying to role my own.

It is a lot of extra code, but I have a flag in place for when we do
upgrade to aws-sdk v3 in the future, an error should be raised in the
test suite to have this patch removed.


(transferred from ManageIQ/manageiq-gems-pending@f3e3e9c54f1e33f9ff046ab5cdc668c0012210d2)