[MiqException] Remove MiqLogFileMountPointMissing
This exception was removed here:

- https://github.com/ManageIQ/manageiq/pull/4131
- https://github.com/ManageIQ/manageiq/commit/5df370256b06f96c8ab2d594ad0b65986fe387b8

and the code for `mount_share` was updated again here:

- https://github.com/ManageIQ/manageiq/pull/5203

This code was not used anywhere else:

https://github.com/search?q=org%3AManageIQ+MiqLogFileMountPointMissing&type=Code

So the one case it was used is unreachable.  Removing it seems like a
reasonable solution.


(transferred from ManageIQ/manageiq-gems-pending@1e483046ec7eed086b1b23f1f1115a8ec82e9efc)