Rescue TypeError During PEheader Processing
Currently TypeError is ignored during processing of the .ico files in the PEheader processing code.
This results in all the information collected about all files in the VM being discarded and no new
info being available about said files.

Rescue the TypeError so that processing of the files can continue successfully.