Address Review Comments About globPattern Variable
Review comments point out that the existing globPattern variable
is only used in the true path of the "if @fs.fileExists?" and as such
its initialization can be moved to the same path.  In addition, it
is not actually treated as a globPattern any longer and as such will
be renamed.