Check Validity of run spec
The run spec LCN and length values need to be checked for nil values
prior to using them in the computation to determine which blocks to read.
If invalid data is read from the virtual disk, an exception
will occur.

One place this impacts is the the calulate_digest method of the MD5deep
class used to collect file information in smartstate.  If this
failure occurs all data is lost.  A rescue clause will be added to
make sure other errors are not hit causing similar issues.