Handle Errors during calculate_digest
If errors reading the disk to calculate the file digest are received,
information such as the file access time will be lost, and not written to
the xml file.  Make sure that calculate_digest handles the error to avoid this.
The upshot of an error like this is that an exception occurs parsing the
time data in the xml when it is supposed to be written to the database.
That exception would cause all scan_profile data to be lost.